Move Application service to application layer
